# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Snipr URL shortener API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /links/{id}/tags:
    get:
      operationId: getLinkTags
      tags: [links]
      summary: Get tags for a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tags for this link
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    put:
      operationId: setLinkTags
      tags: [links]
      summary: Set tags for a link (replaces all existing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tagIds:
                  type: array
                  items:
                    type: string
              required:
                - tagIds
      responses:
        "200":
          description: Tags set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /links/{id}/rules:
    get:
      operationId: getLinkRules
      tags: [rules]
      summary: Get routing rules for a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: setLinkRules
      tags: [rules]
      summary: Set routing rules for a link (replaces all existing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateLinkRuleRequest"
              required:
                - rules
      responses:
        "200":
          description: Rules saved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    LinkRule:
      type: object
      properties:
        id:
          type: string
        linkId:
          type: string
        type:
          type: string
          enum: [geo, device, ab, rotator]
        priority:
          type: number
        conditions:
          type: object
          description: "geo: {countries:[]}, device: {devices:[]}, ab/rotator: {weight:number}"
        destinationUrl:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - linkId
        - type
        - priority
        - conditions
        - destinationUrl
        - createdAt
    CreateLinkRuleRequest:
      type: object
      properties:
        type:
          type: string
          enum: [geo, device, ab, rotator]
        priority:
          type: number
        conditions:
          type: object
        destinationUrl:
          type: string
        label:
          type: string
          nullable: true
      required:
        - type
        - destinationUrl
...
    Tag:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - color
        - createdAt
        - updatedAt
...